Initial protean wrapper
This change adds a protean compiler simulated annealing wrapper as a
step in the clang driver. To invoke protean tool as part of compilation
flow we use the clang flag -OP. For the moment, protean tool is the same
as opt, but that will change shortly.

Added -mllvm -protean flag support to ensure we do not optimize with entire
O3 pipeline when running protean flow.