[Protean] Initial protean wrapper
This change adds a protean compiler simulated annealing wrapper as a
step in the clang driver. To invoke the protean tool as part of compilation
flow, we use the clang flag -OP. For the moment, the protean tool is the
same as opt, but that will change shortly.

Add -mllvm -protean flag support to ensure we do not optimize with the
entire -O3 pipeline when using the protean flow.