Add support for phase ordering + Bug Fixes
This patch implements the Base class for Simulated Annealing and
PhaseOrderGenerator.

Now with --protean -O#, the compiler will set all the parameters
according to -O# and enable the Simulated Annealing for
the phase ordering process.

[Protean] Add support for phase ordering + Bug Fixes

Offering: BiSheng

This patch implements the Base class for Simulated Annealing and
PhaseOrderGenerator.

Now with --protean -O#, the compiler will set all the parameters
according to -O# and enable the Simulated Annealing for
the phase ordering process.