[Protean] Add cost function for simulated annealing
Added a basic cost function to be used in the simulated annealing
process. The user can specify the way the cost will be calculated using
the command line option -cost-type. The cost can be calculated by either
counting instructions in the IR, or by the size of the bitcode file
(-cost-type=instcount or -cost-type=file-size, respectively).