Add random sequence generation to Protean
1) Added a new option for a cost function for the protean code
optimizer, utilizing the llvm-mca tool. To use this to calculate the
cost for Simulated Annealing, specify -cost-type=MCA when invoking
protean.

2) Added random sequence generation to protean. Now, during Simulated
Annealing, we generate all possible permuations given a set of recipes,
shuffle the array according to a set seed, and iterate through that
array for the iterations specified by the user.